move DomainHelper logic to Not.DomainUtils namespace
